Track organisation homepage as a `finding` page in Google Analytics
Add meta tag which marks the organisation homepage as being in the
`finding` (navigation) stage of a user journey, as opposed to the
`thing` (content) stage.

This will enable us to analyse how users navigate the site so that we
can work out whether the new navigation pages are helping users find
what they need.

Most pages use the shared meta tags in static to track this value, but
organisation homepages are not in the content store so this meta tag
must be added separately.