Access Jenkins build parameters with `params` instead of `env`
There are three ways of accessing build parameters in Jenkins pipelines:
`env.FOO`, `FOO` and `params.FOO`. But the first two versions don't work
on the very first build of a branch (see JENKINS-40574).

We have been using `env.FOO` and working around the problem by calling
the `initializeParameters` function. But it is simpler to use
`params.FOO` instead, and it means that the call to
`initializeParameters` is no longer needed.