Only throttle the test step, not the whole CI build
Delete the `ThrottleJobProperty` config from the Jenkinsfile. The
Throttle Concurrent Builds plugin isn't fully compatible with pipeline
jobs. The `maxConcurrentPerNode` setting is ignored, and the
`limitOneJobWithMatchingParams` setting was causing schema test builds
to pile up in an executor queue because all of their parameters matched.

Since each job is executed in its own workspace, it should be safe to
run multiple whitehall builds in parallel. So the throttling parameters
have been removed entirely.

It's not safe to run the tests in parallel though, so this commit adds a
lock which prevents the tests being run on the same node.

This should speed up schema test builds because multiple whitehall
schema tests can be run in parallel.