Add draft links check to DocumentCollection
Now that we have edition links for the `DocumentCollection` format we
can check them in the sync checks.

This commit adds links to the sync check. Edition links will include
`draft` and `withdrawn` documents so the filtering that is applied to
the links checked against live is not implemented for checks against the
draft content store.

There are also Documents linked that have been deleted. This happens at
the edition level so, other than the having their slug changed to
`deleted-` the document doesn't currently know that it has been deleted.
We exclude those from the checks.