Add analytics meta tags to navigation pages
Add meta tag component to pages which are used for navigation:
organisations, policy areas, topical events and the organisations
homepage.

The meta tag component requires a content item, so this change also
loads the content item on each of those pages.

This will help us track users' journeys through the navigation and
content elements of the site to work out whether they find the content
they're looking for.