Request draft taxons to publishing api
In this commit we have hardcoded the content_id of two taxons that we
know are currently in draft state.
The intention is to allow users to tag documents to those two taxons,
for that, we need to make a separate request to Publishing API for each
draft taxon.

History

We thought of making two separate requests to Publishing API, the first
would only have live taxons and the second would have live and draft
taxons. We would then identify the ones not in the first request and
mark them as draft. Although this wasn't ideal because it was looking
into the Education Taxonomy only, that means that any draft taxon that
is not a child of Education wouldn't be considered.

Once the functionality has been tested we can then move it into https://github.com/alphagov/govuk_taxonomy_helpers

Trello: https://trello.com/c/LmVHv7Jv/537-build-design-for-tagging-to-draft-taxons-in-whitehall