Add People Payload Builder and tests
This payload builder takes an extra argument in the form of a symbol
which should specify the required key for use in the Links hash.

This will allow us to use this payload builder to create both the
`people` and `ministers` arrays in the Links hash for formats that
require them.

Both `people` and `ministers` will contain the same information at this
point. The duplication of data at this point makes it easier for
services downstream to consume the data.