Remove Ministers PayloadBuilder
The functionality provided by `PayloadBuilder::Ministers` is now
provided by `PayloadBuilder::People`. A parameter in `People#for` allows
us to specify the key for which the array should be created. In the case
of ministers, the call would look like:

```
PayloadBuilder::People.for(edition, :ministers)
```

The payload builder for People is also used when creating the `people`
array in the links hash for email-alert-api.