Fix routes helper test
Fixes a test that was previously using `Speech` as an example of an
unmigrated format. Switching the rendering app for `Speech` now updates
its preview-url to use draft origin.

This commit updates the test to use the last non-migrated edition
(`CorporateInfoPage`). The test can be removed when that format is
migrated.