Skip virus checks for attachments in development
The attachments that have been uploaded for documents in production
do not exist on our local development environment. Since these files
don’t exist, they don’t ever get moved to the `clean_path` (even with
85b98996ee86ad3b28e5bfa89f03c840bc32b3d6). Therefore, it looks like
there is a permanent virus check pending for all attachments, and it
isn’t possible to republish an existing document with attachments.

This change adds a check for the development environment, and whether
the file actually exists (hence retaining compatibility with the fake
virus checker implemented in the above `govuk-puppet` commit).