Fix GOVUK_APP_DOMAIN
This is defined in a few places and was using `alphagov.co.uk`. This was
causing inconsistencies depending on how tests were run. It is redefined
in the Rakefile so tests run with rake were picking up different values
from Plek than tests run directly. This resulted in tests that would
fail when run individually in the VM but pass on CI and vice versa if
the test was updated to pass locally.

This commit removes some of the definitions and updates the default
value to `test.gov.uk`. It also updates tests and factories to use the
new URL.