Ensure "How Government Works" works
Whatever the outcome of the election, there will be a short period
where the government is being changed over when all ministerial roles
are unallocated. Previously that would cause the How Government Works
page to error, because it always looks for the current PM's picture.

During the last election we temporarily hard-coded the picture to be
a shot of the Number 10 door. This change restores that picture, and
uses it when there is no currently allocated PM.