Add SVG support
This allows the SVG file format to be uploaded and rendered via
Whitehall frontend.

SVG paths are not currently being sent to publishing API as they are
generally being generated for a specific size and this returns `nil` for
SVG as they don't have thumbnails. This will be fixed in subsequent
commits.