Fix multiple image upload script and test
This commit fixes the issue where the input fields of cloned image form elements were not cleared. This made it impossible for a user to save an edition when dealing with invalid images (as choosing a new valid image would create a new image form field with the settings of the invalid image). The problem was the JQuery's `children` method only looks at an element's immediate children, but the elements are more deeply nested. Replacing it with the `find` method makes it work as intended.

The JS test for this feature also appeared to be quite out of date. The commit refactors the test fixture to be a block of HTML rather than appending to an element programmatically, which matches the fixture lower down in the code and makes it easier to see the HTML structure at glance (which also helps to compare it to the view template). It also updates ids and names of the elements to match the rendered form. Some of the elements in the test fixture do not get rendered anymore (title, accessible), so we replace them with the new field (alt_text).

Rendered Image form element (from Production) used for reference to create the updated test fixture:
```
<fieldset id="image_fields" class="images multiple_file_uploads">
  <legend>Images</legend>

  <div class="file_upload well">
    <h3 class="remove-top-margin">New image</h3>

    <div class="form-group">
      <label for="edition_images_attributes_0_image_data_attributes_file">File</label>
      <input class="js-upload-image-input" type="file" name="edition[images_attributes][0][image_data_attributes][file]" id="edition_images_attributes_0_image_data_attributes_file" />
      <input type="hidden" name="edition[images_attributes][0][image_data_attributes][file_cache]" id="edition_images_attributes_0_image_data_attributes_file_cache" />
    </div>

    <div class="form-group">
      <label for="edition_images_attributes_0_alt_text">Alt text</label>
      <input class="form-control" type="text" name="edition[images_attributes][0][alt_text]" id="edition_images_attributes_0_alt_text" />
    </div>

    <div class="form-group">
      <label for="edition_images_attributes_0_caption">Caption</label>
      <textarea rows="2" class="form-control" name="edition[images_attributes][0][caption]" id="edition_images_attributes_0_caption">
      </textarea>
    </div>
  </div>
</fieldset>
```