Clear the caption textarea on image form cloning
The caption textarea should also be cleared when cloning the image form element, previously only the input fields would be reset.

Note that for some reason, using `.val('some-value')` to set the value in the test does not work properly here in that when the element is cloned, the textarea is already empty. Instead it should be set the value of the cloned textarea in the test. Using `append` does the right thing so we use that instead.