Handle previews of CSVs with mixed newlines
The ruby CSV library tries to automatically detect the row separator used
in a file that it opens.  It does so by scanning ahead for the first
newline sequence it can find in the data and assuming that this is the row
separator.  What this means is that an embedded newline in a cell on the
first row can cause problems if it is a different sequence to that used in
the rest of the file.

See the description of `row_sep` in the ruby-doc for CSV.new for an
explanation[1].

We've occasionally seen files with this problem so we can try to mitigate
against it.  We try to open the file and read the first row.  If this
raises a MalformedCSV error we try to open the file again with an explicit
`row_sep` of `\r\n`.  If this still raises a MalformedCSV error we raise
the original version.

We've only seen an embedded `\n` in an otherwise `\r\n` file so that's all
we try to resolve the problem.  We might want to add extra retries to cycle
through all possible newline sequences.

Note that we have to edit test/fixtures/malformed.csv to make it even more
malformed.  Without editing this file can now be handled by the CsvPreview
as it exhibits the exact problem we've mitigated.  We make it more broken
by adding an extra `\r` to one of the line endings.
[1]: http://ruby-doc.org/stdlib-2.2.3/libdoc/csv/rdoc/CSV.html#method-c-new