Remove world edition rules
These rules were always returning true and as indicated by the comment, are therefore redundant.
This was added in 2013 in expectation of more complex rules around the WorldLocationNewsArticle format.
It hasn't changed since then so we remove it.
We also remove the tests for these rules. WorldLocationNewsArticles rules still get tested by the more generic
roles tests which test permissions for Editions, which WorldLocationNewsArticles inherit from.