Update test gems
This commit removes version locking for most gems used for testing to upgrade them to the latest versions and keep them up-to-date. It also locks the versions of some production gems to prevent them being upgraded for the time being. Gems related to parallel testing are version-locked due to incompatibilities with new versions and the whitehall parallel testing setup.