Convert 'foreign language only' from link to checkbox
The main change here is that instead of a link we now use a checkbox to trigger the showing of the foreign language options. The HTML structure and UX design is copied from the ‘schedule publishing’ further down which uses the same pattern. This also means that we do not show/hide the fieldset but instead only the div.foreign-language within the fields.

In the template we check the `primary_locale` of the edition so that we can determine whether the checkbox should be checked after an invalid form submission.