Add `.search_filter_types` to `NewsArticleType`
As part of the process of migrating the `WorldLocationNewsArticle`
format we need to be able to include/exclude them from the announcement
filter (at `/government/announcements`). As they will be a subtype of
`NewsArticle` we need the ability to return the `search_format_types` of
all subtypes to enable this functionality. Currently the filter just
supplies `news-article` to rummager which returns all sub-types.

This commit adds a class method to `NewsArticleType` that returns an
array of `search_format_types` for all types.