Refactor admin edition form js test
One of the problems with the fixture in this file is that it’s very generic and does not represent any specific page in Whitehall. We attempt to split up the fixture into separate parts so that the tests can load only the HTML fixtures that they are testing and to make the fixture more adaptable and configurable. This is in preparation for adding tests specific to News Articles.