Toggle and reset world news story fields
This commits adds JS to toggle visibility and reset fields based on the selected ‘News Article’ type. To help with selecting the DOM elements, we add specific class names to the existing fieldsets and add fieldsets to elements that did not have them.

Working the JQuery Chzn select fields was a bit fiddly. The main realisation was that for clearing out select fields that support ‘multiple’ selected options you had to use `.val({})` (rather than `.val(‘’)` as used for single selection fields). We also only clear the field when a value is set, otherwise it leads to weirdness such as truncated placeholder text…