Fixes issue preventing draft taxon tree resolution
Owing to poor default arguments in a private method, Whitehall doesn't
query publishing-api for the draft children of a taxon. This prevented
it from rendering a full taxonomy tree for draft root taxons.

This only became an issue today, when a full draft taxonomy was released
to editors for tagging.