Fix breadcrumb preview box for new taxonomy
Made childless draft taxons appear in preview box by giving them the
"topic-tree" class.

Gave each taxon a parent_node as we recurse through the taxons when
building the Taxonomy Tree. This is necessary in order to give the
taxon checkboxes the correct "data-ancestors" property for the breadcrumb
preview box.

Trello:
https://trello.com/c/ihpJNOIU/99-whitehall-breadcrumb-preview-for-selected-topics-stopped-working-a-while-ago