Fix person_helper bug
Many of the `Person names` in the Whitehall DB have a trailing space. The
`PersonHelper` generates a list of person names for select options and
where there is more than one person with the same name it disambiguates
the name by appending the persons role.

This fixes a bug where if there are two people with the same name but
one has a trailing space and one doesn't the disambiguation the names
would not be disambiguated and only one of the users would appear in the
select options in the admin UI.

`Person#forename` and `Person#surname` aren't required and one or the
other is nil throughout the test suite. There is currently one person in
the DB that has a nil forename so `.try` has been used to handle this
questionable validation.