Add WLNA -> NewsArticle rake task
We no longer use the `WorldLocationNewsArticle` format. This commit adds a rake
task that converts existing documents of this format to `NewsArticle`

`Unpublishing` objects associated with `WorldLocationNewsArticle` also
have a `#document_type`. This rake task also updates the
`Unpublishing`s when the `Edition` type is changed during the migration
of this format to the `NewsArticle` format.

The rake tasks takes two arguments which indicate the start and end of
the range of document ids that we want to convert. This will allow us to
convert them in batches as it is difficult to debug in integration as
the sync checks require the WLNAs be migrated and then all of the other
NewsArticles to be republished.