Enable translation links for A/B test locations
This commit enables the translation picker for A/B test world locations that have translations. The translated content is rendered in the existing template and the user is re-assigned to the “A” bucket to prevent the confusion of switching between the “A” and “B” layouts of the page depending on the language.