Add route for embassy contact page for A/B test
Include a new method in the WorldwideOrganisations controller to handle
the route and a view. Placing the view under the `embassies` directory
since we are in fact `show`ing an individual embassy.

Using the WorldwideOganisation contoller (instead of the Embassy
controller)  here because we will need some
of the private methods that are there already and it didn't seem worth
it for A/B test to refactor those into a common class. At the same time
duplicating the code felt the wrong thing to do.