Fix incorrect rendering of multiple same Id
Current use of content_tag_for in the worldwide_organisation partial is
causing html to be rendered that has multiple Ids that are the same on a
single page. This is incorrect.

Remove content_tag_for and replace it with a hand coded div.
Drop the use of `id` in the div.