Display available translations only when available
Translation links on the worldwide publishing taxon pages should only
be displayed if there multpile translations. Currently, the
`available_translations` div displays even when the world location is
only available in English.

Fix this to only display the `available_languages` div if there are
multiple languages available for the world location.