Assert that correct links are included
We currently check the value of the `links` by comparing the exact
values. Because we are adding a new key to every item, all the tests
will start failing. To prevent shotgun surgery across the presenter
tests this commit uses the new test helper to just check if the
expected links are included in the generated payload. This means we can
add other keys to the payload without having to change all the tests.