Return AB test link for WorldwideOrganisation
For the worldwide AB test we require that the usual worldwide
organisation path for embassies that are part of the test redirect to
the parent location.

When viewing Corporate Information Pages, the header logo is sometimes
a hyperlink back to the Worldwide Organisation. If we are in the B
cohort, we want this link to take us to the new (B cohort) Worldwide
Organisation page.