Return AB test path for WorldwideOrganisation
For the worldwide AB test we require that the usual worldwide
organisation path for embassies that are part of the test redirect to
the parent location.

In certain circumstances we still need to link to the embassy from pages
e.g. corporate information pages and this requires that the links are
updated with the new path to avoid the redirect and go to the new
embassy pages.

This commit temporarily overrides the path and url helpers for
`WorldwideOrganisation` to achieve this for the duration of the test.