Guard against non-testable object arguments in `is_under_test?`
The changes to the path/url helpers have caused a test or ten to fail
where the helpers are used but the tests don't have locations set up on
the organisations under test. Some other tests provide a `String` for
locatin.

As this code is currently for use during the AB test I've added a guard
for object arguments that don't respond to slug to the
`WorldwideAbTestHelper#is_under_test?` method that returns false in these
cases. This is preferable here rather than add even more object creation
to the already slow tests which would also further spread A/B test code
around the application.