Add test for email signup for News articles
News articles can be associated to WorldLocations and emails do already get sent when a user is signed up to a WorldLocation and a NewsArticle (for any subtype, including 'World news story') for that location is published. This commit adds a feature test for that. Note that the first_published date needs to be set for the test to pass, otherwise the `published_today?` method in the `GovUkDelivery::Notifier` class would fail.