Add an election task to clear ministerial roles
In the event that there is a continuation of current government
following the election it may be necessary to clear the role
appointments, with the exception of the prime ministerial role, to allow
the content team to rebuild the ministerial list without ending the
current government.

This commit adds a rake task to accomplish this.