Send the users organisation slug to GA as a custom dimension
The performance analysts want to know which departments are using the
publishing applications and have agreed that sending to GA the slug of the
organisation that each logged in user belongs to is the best way to
achieve this.  If the user does not have an organisation we send `(not
set)` instead.  We set it as custom dimension 8 post once they're logged
in so it will be sent along with the track page view.

Not entirely happy with the way we test this, we end up testing that the
JS code that would trigger the GA event is on the page, not that the event
is actually triggered.