Add organisations eligible for world tagging
Add FCO, UKVI, DIT, DFID, DExEU to the list of organisations allowed to
tag documents in Whitehall admin.

The YAML file is split into `education_related` and `world_related` to
which will help keep logic neat when we need to selectively display the
tagging interface based on Document Type and Organisation combination.