Make world location base_path less
WorldLocation pages are being converted to worldwide taxonomy and will
no longer be renderable. To facilitate this we will now consider them
base_path-less content.

This commit updates the schema for this format and modifies the
presenter to match.

WorldLocations of type InternationalDelegation will continue to be
rendered by Whitehall so the presenter will still present a base_path
for those to the publishing API.