Add summary to WorldwideOrganisations presenter
WorldwideOrganisations should display a summary when they are viewed
through the new taxonomy navigation pages. Including the summary in the
'description' field in the placeholder present should allow the
`collections` application to correctly display the summary.

Also ensure that description is sent to rummager with the value of the
summary.