Fix an over-specific scope joining StatisticsAnnouncements to Publications
The StatisticsAnnouncements.without_published_publication scope was failing
to include announcements associated with deleted publications, because
the 'publication_id' referenced in the `where` clause is not NULL in that
case.

By changing the WHERE clause to reference the `id` field of the `included`
table, ActiveRecord builds a left-outer join which gives us the resultset
we expect.