Rake task to add /world prefix route to router
This commit adds a rake task, which adds a `/world` prefix route to the router. This route will overwrite the exising prefix route which initiates a redirect inside the router. This new prefix route tells the router to send all requests to `/world/*` to whitehall, which will then serve a redirect to `/government/world`.

This is in preparation for the move from `/government/world` to `/world`.