Test saving World Location calls the News Page Worker
When saving a World Location, we must also publish the World Location
News Page to PublishingApi and send it to Rummager. This is done by the
World Location News Page Worker. Test that on saving a World Location,
the Worker is called.

Also stub requests to World Location News Page Worker in other tests
that save World Location.