Remove translation dependent world location features
We are removing support for translated world location content and the
world location pages in general as they are becoming taxons.

This commit removes some failing features to allow us to deploy the code
required to move the world location over to the taxon pages.

Another PR will remove the other world location code that is no longer
used once we have switched them over.