Allow uploads of shp_rxl files
When uploading zip files containing shape files, a user reported the
system didn't allow. After looking, I realised one of the files with a
`.shp_rxl` extension isn't in our list of allowed extensions.

This commit adds it to the optional list of allowed extensions so that
we allow uploads.