Show World Location News Translation
Using the new World Location News Translation model, show the available
translations for World Location News. Note that the World Location News
Translation model is a very simple model to proxy the
`translated_locales` from a World Location through to the front end.

We have done this because World Locations are now served by Collections
and set up as Taxonomies in Content-Tagger. Whitehall is sending World
Locations as base pathless content items to Publishing Api. We don't
actually want to send the various World Location translations to
Publishing Api anymore since World Locations themeselves are not being
translated. However, we do need to allow users to edit the translations
for World Locations in Whitehall because the mission statement is
displayed on the World Location News page.