Redirect non-en world locations
We no longer support translated world location pages. They were
redirected but re-appeared when we moved them from
`/government/world/<slug>` to `/world/<slug>`.

This commit adds a rake task to redirect them to their :en equivalent
again.

This happens in two steps. Firstly, we update the publishing API to
change the base_path of the document to `/world/<slug>.<locale>`. Then
we make the `unpublish` call with the redirect. We can't just redirect
via publishing API as that would create a redirect from the original url
not the new `/world` one.

As the presenter no longer exposes a base path for WorldLocations we
can't use the existing worker classes for publishing. This commit adds a
worker which reinstates the base path by adding it to the payload from
the presenter. This worker can be removed once this rake task has been
run.