Remove locale patch from WorldLocation
This was patched to allow us to remove the translations whilst migrating
world locations to a taxonomy. Unofortunately we need translations
present on this object to allow featuring of news articles on the
translated news pages.

This commit exposes available and translated locales again. A subsequent
commit will handle stopping them being published as content items.