Only publish :en WorldLocation
The world news page functionality bases the translations it supports and
their featured article lists on the WorldLocation object. We are no
longer rendering world locations themselves and we only want to send the
:en version to the publishing API.

This commit overrides the `PublishesToPublishingApi` functionality to
only send the :en version whilst maintaining the translations within
Whitehall for use by the news pages.