Discard drafts of deleted HtmlAttachments
This fixes a bug where if an HtmlAttachment is deleted and then the
draft edition that it is attached to is discarded the publishing API was not
being updated so the draft of the attachment would be left in the
content store. This causes problems if the publisher then tries to
recreate the document and attachment as the draft blocks the new
attachment from being published and as it occupies the base_path the new
edition will appear to be linking to the previous version.