Fix duplicate content_id bug
When an `HtmlAttachment` is created on an `Edition` we check previous
`Edition`s to see if an attachment with the same title (and
consequently the same `base_path`) has existed before and if it does we
reuse the `content_id` otherwise we get a conflict when the most recent
`HtmlAttachment` is sent to the publishing API. Previously deleted
`HtmlAttachment` are redirected to the parent document so they still
have a representation in the publishing API (albeit a redirect but it
still causes a conflict).

This was being done by comparing the title of the new attachment with
all titles of attachments that have been associated with editions of the
parent document.

It is possible to rename `HtmlAttachment`s though which means that the
objects `title` and `slug` aren't always related. This can cause very
weird redirecting behaviour.

e.g.

'Edition one' (slug `edition-one`) gets an attachment 'Attachment one'
(slug `attachment-one`). So the attachment base_path is
`/edition-one/attachment-one`.

The attachment is then renamed to 'Attachment one rename'. Its slug
stays the same.

We then add another attachment called 'Attachment one rename' which gets
the slug 'attachment-one-rename'. The current code compares the titles
and takes the `content_id` from the original attachment. We now have two
attachments with the same title, same content_id but different slugs.

Depending on which attachment makes it to publishing API first we get a
variety of different combinations of strange redirects from one to the
other.

This commit compares using the slug rather than the title. The problem
with this approach is that the slug and the content_id are both created
in callbacks so we now need to make sure the slug is created before the
content_id. This has involved taking the `HasContentId` module out of
the base class.

Unfortunately translated `HtmlAttachment`s don't have slugs
(agghhhhh) so this approach won't work for them. The previous bug ridden
title comparison has been maintained for them until inspiration strikes.