Remove the UK world location
This commit removes the UK world location and untags all editions that are tagged to it.

It doesn’t make sense to have a world location for the UK since all documents are assumed to be about the UK unless expressely tagged to another world location. This also causes confusion for users who sign up to email alerts for the UK world location but are not alerted to the majority of changes since tagging is inconsistent.