Populate the analytics meta tags for finders
This commit makes sure we set the `@content_item` instance variable with
the content store response related to the finder the user is on, so that
the analytics meta tag component gets populated correctly and all
necessary meta tags added to the page.

This will allow us to identify what page we are on in static and
simplify the code that looks for the number of links and sections on the
page.