Link to live site for published html attachments
Whitehall Admin UI was still linking to whitehall-frontend for html attachments for published documents, so this commit links them to the live site instead.

We also reworked the HtmlAttachment `url` method a bit so that it returns either the full URL or just the relative path, based on context.

The `HtmlAttachment::url` method is being used in two contexts:
1. When rendering links on the admin edition show page
2. When creating the payload via the PublishingApiHtmlAttachmentPresenter

In case of 1. we need the full URL but for case 2. we only need the relative path. This
commit enables that with an extra option param. The method will default to pick the relative path.