Fix from/to date filtering in whitehall-admin
This commit changes the from/to date filtering logic for whitehall-admin to correctly include the actual from date in the range when filtering documents, and not return some documents after the "to" date. This logic was previously only returning documents last updated after/before 12 noon. It also changes the heading that appears to make it clearer that the from date is included in the range.