Upgrade qunit and sinon versions used in JS tests
We go up to qunit 1.20.0 and sinon 1.17.2.  These specific versions are
used because they're the ones used by a gem we want to replace test_track
with.  Unfortunately these are also quite out of date, but they'll do for
now until we explore how to upgrade the versions we use with the other
gem.

We've had to change some of the tests to work with the new qunit and sinon
versions.  Notably we've used `sinon.test` methods around most of the tests
that use stubs and mocks as this means we don't need to manually restore
the stubbed and mocked functions to their original functionality.  We also
need to change the phantom-driver to use callbacks instead of defining
functions when hooking into qunit.