Add shim for history API for phantomjs
With the new version of sinon we're no longer able to stub the history
pushState and replaceState methods directly.  Doing so fails with

   Attempted to wrap undefined property pushState as function

Errors from sinon.  It's not 100% clear if phantomjs (particularly the
version we use) actually supports the history API or not, but adding
this tiny shim is enough to let us stub the part of the API we want
to stub in the tests where we want to do that.  This shim is absolutely
not intended as a full implementation of the history API, just enough to
let our tests pass.