Replace test_track with teaspoon
The test_track gem does not play nice with newer sprockets-rails versions
and is a blocker to us upgrading.  Teaspoon is an (almost) drop in
replacement that lets us upgrade.  While it ships with qunit and sinon
it doesn't ship with sinon-qunit which we need to tie the two libraries
together.