Make EmailTopicChecker use the /topic-matches endpoint
The previous approach of retrieving topics from
email-alert-api was wrong. Instead of using the
same links/tags from UrlToSubscriberListCritera,
we need to replicate the behaviour that happens
downstream which means presenting the edition
with PublishingApiPresenters.

email-alert-api didn't provide a mechanism to
run match queries directly, so a new endpoint
was added '/topic-matches' that provides this
functionality.