Do not send empty speech type explanation
Of the seven SpeechTypes, only three of them have an explanation string set. For the speech types without an explanation, the presenter was sending speech_type_explanation as `nil` which failed against the content schemas. Also government frontend only checks for presence of the key, so sending an empty string would cause it to render an empty pair of parentheses. See code in government-frontend below:

```
  def speech_type_explanation
    explanation = content_item["details"]["speech_type_explanation"]
    " (#{explanation})" if explanation
  end
```

Therefore not sending the key for speech_type_explanation at all when empty seems like the easiest fix for this. This should also make the Speech sync checks perform better.