Pass ‘supertypes’ to /topic-matches endpoint
The SubscriberListQuery in email-alert-api can
return additional topics based on the supertypes
of the edition. We want to make sure we provide
this endpoint with these supertypes so we get an
accurate view of which topics are matched.

This commit also includes additional output to
show which requests are made to email-alert-api
and govuk-delivery.