Remove version constraint from sprockets-rails and update
We're now on v 3.2 which means we need to change some of the configuration
in the app relating to assets.  We move precompile, prefix, and version
details into the assets intializer so it's more like a fresh rails 4 app.

We also bump shared_mustache to 1.0.1 so that we avoid an issue with the
assets it uses not being declared for precompilation.