Move sendfile header config to production env
This caused lots of

    X-Accel-Mapping header missing

log messages to appear when running tests because Rack expects nginx to
forward it when using sendfile, but we're not running inside nginx for
our tests so this header isn't present.  The default location in a new
rails app for setting this behaviour is the production environment file
so that's where we move it and avoid the log output.