Add rake task to redirect world locations to their taxon pages
We already route any '/government/world/*' page to '/world/*' via Whitehall's routing. This updates the publishing api (and content store) to do the same. Without this, we're seeing odd effects like World Locations showing up twice in Search. We also remove the duplicate entries from Rummager.

Initially we tried just unpublishing the existing WorldLocations from within Whitehall via the Unpublish action, but that didn’t work because PublishingAPI would error since the RedirectsAndRoutesValidator needs a base_path. Then we tried creating a migration in PublishingAPI which temporarily added the base_path back in. This worked but showed that on unpublishing and redirecting all the ‘/government/world/*’ content would also untag any content from it (eg any news articles tagged to a country would suddenly be untagged).

After chatting with @kevindew about this, we finally settled on this current solution which publishes Redirect items for those base_paths (but with different content_ids) and places redirect routes into the router database. This solution still allows us to retain the tagging to WorldLocation content items in the PublishingAPI.